migrating fee recipient endpoints to pure http implementation
