+++ validator/rpc/apimiddleware/endpoint_factory.go
@@ -17,7 +17,6 @@ func (f *ValidatorEndpointFactory) IsNil() bool {
 func (*ValidatorEndpointFactory) Paths() []string {
 	return []string{
 		"/eth/v1/keystores",
-		"/eth/v1/validator/{pubkey}/feerecipient",
 		"/eth/v1/validator/{pubkey}/gas_limit",
 	}
 }
@@ -32,10 +31,6 @@ func (*ValidatorEndpointFactory) Create(path string) (*apimiddleware.Endpoint, e
 		endpoint.PostResponse = &ImportKeystoresResponseJson{}
 		endpoint.DeleteRequest = &DeleteKeystoresRequestJson{}
 		endpoint.DeleteResponse = &DeleteKeystoresResponseJson{}
-	case "/eth/v1/validator/{pubkey}/feerecipient":
-		endpoint.GetResponse = &GetFeeRecipientByPubkeyResponseJson{}
-		endpoint.PostRequest = &SetFeeRecipientByPubkeyRequestJson{}
-		endpoint.DeleteRequest = &DeleteFeeRecipientByPubkeyRequestJson{}
 	case "/eth/v1/validator/{pubkey}/gas_limit":
 		endpoint.GetResponse = &GetGasLimitResponseJson{}
 		endpoint.PostRequest = &SetGasLimitRequestJson{}
