+++ validator/rpc/standard_api_test.go
@@ -4,11 +4,9 @@ import (
 	"context"
 	"encoding/hex"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"testing"
 
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/golang/mock/gomock"
 	"github.com/golang/protobuf/ptypes/empty"
@@ -540,455 +538,6 @@ func createRandomKeystore(t testing.TB, password string) *keymanager.Keystore {
 		Version:     encryptor.Version(),
 		Description: encryptor.Name(),
 	}
-}
-
-func TestServer_ListFeeRecipientByPubkey(t *testing.T) {
-	ctx := context.Background()
-	byteval, err := hexutil.Decode("0xaf2e7ba294e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2c06bd3713cb442072ae591493")
-	require.NoError(t, err)
-
-	type want struct {
-		EthAddress string
-	}
-
-	tests := []struct {
-		name   string
-		args   *validatorserviceconfig.ProposerSettings
-		want   *want
-		cached *eth.FeeRecipientByPubKeyResponse
-	}{
-		{
-			name: "ProposerSettings.ProposeConfig.FeeRecipientConfig defined for pubkey (and ProposerSettings.DefaultConfig.FeeRecipientConfig defined)",
-			args: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(byteval): {
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9"),
-						},
-					},
-				},
-				DefaultConfig: &validatorserviceconfig.ProposerOption{
-					FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-						FeeRecipient: common.HexToAddress("0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF"),
-					},
-				},
-			},
-			want: &want{
-				EthAddress: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			},
-		},
-		{
-			name: "ProposerSettings.ProposeConfig.FeeRecipientConfig NOT defined for pubkey and ProposerSettings.DefaultConfig.FeeRecipientConfig defined",
-			args: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{},
-				DefaultConfig: &validatorserviceconfig.ProposerOption{
-					FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-						FeeRecipient: common.HexToAddress("0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9"),
-					},
-				},
-			},
-			want: &want{
-				EthAddress: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			},
-		},
-		{
-			name: "ProposerSettings is nil and beacon node response is correct",
-			args: nil,
-			want: &want{
-				EthAddress: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			},
-			cached: &eth.FeeRecipientByPubKeyResponse{
-				FeeRecipient: common.HexToAddress("0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9").Bytes(),
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctrl := gomock.NewController(t)
-			mockValidatorClient := validatormock.NewMockValidatorClient(ctrl)
-
-			m := &mock.MockValidator{}
-			err := m.SetProposerSettings(ctx, tt.args)
-			require.NoError(t, err)
-
-			if tt.args == nil {
-				mockValidatorClient.EXPECT().GetFeeRecipientByPubKey(gomock.Any(), gomock.Any()).Return(tt.cached, nil)
-			}
-
-			vs, err := client.NewValidatorService(ctx, &client.Config{
-				Validator: m,
-			})
-			require.NoError(t, err)
-
-			s := &Server{
-				validatorService:          vs,
-				beaconNodeValidatorClient: mockValidatorClient,
-			}
-
-			got, err := s.ListFeeRecipientByPubkey(ctx, &ethpbservice.PubkeyRequest{Pubkey: byteval})
-			require.NoError(t, err)
-
-			assert.Equal(t, tt.want.EthAddress, common.BytesToAddress(got.Data.Ethaddress).Hex())
-		})
-	}
-}
-
-func TestServer_ListFeeRecipientByPubKey_BeaconNodeError(t *testing.T) {
-	ctx := context.Background()
-	byteval, err := hexutil.Decode("0xaf2e7ba294e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2c06bd3713cb442072ae591493")
-	require.NoError(t, err)
-
-	ctrl := gomock.NewController(t)
-	mockValidatorClient := validatormock.NewMockValidatorClient(ctrl)
-
-	mockValidatorClient.EXPECT().GetFeeRecipientByPubKey(gomock.Any(), gomock.Any()).Return(nil, errors.New("custom error"))
-
-	vs, err := client.NewValidatorService(ctx, &client.Config{
-		Validator: &mock.MockValidator{},
-	})
-	require.NoError(t, err)
-
-	s := &Server{
-		validatorService:          vs,
-		beaconNodeValidatorClient: mockValidatorClient,
-	}
-
-	_, err = s.ListFeeRecipientByPubkey(ctx, &ethpbservice.PubkeyRequest{Pubkey: byteval})
-	require.ErrorContains(t, "Failed to retrieve default fee recipient from beacon node", err)
-}
-
-func TestServer_ListFeeRecipientByPubKey_NoFeeRecipientSet(t *testing.T) {
-	ctx := context.Background()
-	byteval, err := hexutil.Decode("0xaf2e7ba294e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2c06bd3713cb442072ae591493")
-	require.NoError(t, err)
-
-	ctrl := gomock.NewController(t)
-	mockValidatorClient := validatormock.NewMockValidatorClient(ctrl)
-
-	mockValidatorClient.EXPECT().GetFeeRecipientByPubKey(gomock.Any(), gomock.Any()).Return(nil, nil)
-
-	vs, err := client.NewValidatorService(ctx, &client.Config{
-		Validator: &mock.MockValidator{},
-	})
-	require.NoError(t, err)
-
-	s := &Server{
-		validatorService:          vs,
-		beaconNodeValidatorClient: mockValidatorClient,
-	}
-
-	_, err = s.ListFeeRecipientByPubkey(ctx, &ethpbservice.PubkeyRequest{Pubkey: byteval})
-	require.ErrorContains(t, "No fee recipient set", err)
-}
-
-func TestServer_ListFeeRecipientByPubkey_ValidatorServiceNil(t *testing.T) {
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-
-	s := &Server{}
-
-	_, err := s.ListFeeRecipientByPubkey(ctx, nil)
-	require.ErrorContains(t, "Validator service not ready", err)
-}
-
-func TestServer_ListFeeRecipientByPubkey_InvalidPubKey(t *testing.T) {
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-	s := &Server{
-		validatorService: &client.ValidatorService{},
-	}
-
-	req := &ethpbservice.PubkeyRequest{
-		Pubkey: []byte{},
-	}
-
-	_, err := s.ListFeeRecipientByPubkey(ctx, req)
-	require.ErrorContains(t, "not a valid bls public key", err)
-}
-
-func TestServer_FeeRecipientByPubkey(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	beaconClient := validatormock.NewMockValidatorClient(ctrl)
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-
-	byteval, err := hexutil.Decode("0xaf2e7ba294e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2c06bd3713cb442072ae591493")
-	require.NoError(t, err)
-
-	type want struct {
-		valEthAddress     string
-		defaultEthaddress string
-	}
-	type beaconResp struct {
-		resp  *eth.FeeRecipientByPubKeyResponse
-		error error
-	}
-	tests := []struct {
-		name             string
-		args             string
-		proposerSettings *validatorserviceconfig.ProposerSettings
-		want             *want
-		wantErr          bool
-		beaconReturn     *beaconResp
-	}{
-		{
-			name:             "ProposerSetting is nil",
-			args:             "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			proposerSettings: nil,
-			want: &want{
-				valEthAddress: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			},
-			wantErr: false,
-			beaconReturn: &beaconResp{
-				resp:  nil,
-				error: nil,
-			},
-		},
-		{
-			name: "ProposerSetting.ProposeConfig is nil",
-			args: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: nil,
-			},
-			want: &want{
-				valEthAddress: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			},
-			wantErr: false,
-			beaconReturn: &beaconResp{
-				resp:  nil,
-				error: nil,
-			},
-		},
-		{
-			name: "ProposerSetting.ProposeConfig is nil AND ProposerSetting.Defaultconfig is defined",
-			args: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: nil,
-				DefaultConfig: &validatorserviceconfig.ProposerOption{},
-			},
-			want: &want{
-				valEthAddress: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			},
-			wantErr: false,
-			beaconReturn: &beaconResp{
-				resp:  nil,
-				error: nil,
-			},
-		},
-		{
-			name: "ProposerSetting.ProposeConfig is defined for pubkey",
-			args: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(byteval): {},
-				},
-			},
-			want: &want{
-				valEthAddress: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			},
-			wantErr: false,
-			beaconReturn: &beaconResp{
-				resp:  nil,
-				error: nil,
-			},
-		},
-		{
-			name: "ProposerSetting.ProposeConfig not defined for pubkey",
-			args: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{},
-			},
-			want: &want{
-				valEthAddress: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			},
-			wantErr: false,
-			beaconReturn: &beaconResp{
-				resp:  nil,
-				error: nil,
-			},
-		},
-		{
-			name: "ProposerSetting.ProposeConfig is nil for pubkey",
-			args: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(byteval): nil,
-				},
-			},
-			want: &want{
-				valEthAddress: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			},
-			wantErr: false,
-			beaconReturn: &beaconResp{
-				resp:  nil,
-				error: nil,
-			},
-		},
-		{
-			name: "ProposerSetting.ProposeConfig is nil for pubkey AND DefaultConfig is not nil",
-			args: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(byteval): nil,
-				},
-				DefaultConfig: &validatorserviceconfig.ProposerOption{},
-			},
-			want: &want{
-				valEthAddress: "0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9",
-			},
-			wantErr: false,
-			beaconReturn: &beaconResp{
-				resp:  nil,
-				error: nil,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			m := &mock.MockValidator{}
-			err := m.SetProposerSettings(ctx, tt.proposerSettings)
-			require.NoError(t, err)
-			validatorDB := dbtest.SetupDB(t, [][fieldparams.BLSPubkeyLength]byte{})
-
-			// save a default here
-			vs, err := client.NewValidatorService(ctx, &client.Config{
-				Validator: m,
-				ValDB:     validatorDB,
-			})
-			require.NoError(t, err)
-			s := &Server{
-				validatorService:          vs,
-				beaconNodeValidatorClient: beaconClient,
-				valDB:                     validatorDB,
-			}
-
-			_, err = s.SetFeeRecipientByPubkey(ctx, &ethpbservice.SetFeeRecipientByPubkeyRequest{Pubkey: byteval, Ethaddress: common.HexToAddress(tt.args).Bytes()})
-			require.NoError(t, err)
-
-			assert.Equal(t, tt.want.valEthAddress, s.validatorService.ProposerSettings().ProposeConfig[bytesutil.ToBytes48(byteval)].FeeRecipientConfig.FeeRecipient.Hex())
-		})
-	}
-}
-
-func TestServer_SetFeeRecipientByPubkey_ValidatorServiceNil(t *testing.T) {
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-
-	s := &Server{}
-
-	_, err := s.SetFeeRecipientByPubkey(ctx, nil)
-	require.ErrorContains(t, "Validator service not ready", err)
-}
-
-func TestServer_SetFeeRecipientByPubkey_InvalidPubKey(t *testing.T) {
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-	s := &Server{
-		validatorService: &client.ValidatorService{},
-	}
-
-	req := &ethpbservice.SetFeeRecipientByPubkeyRequest{
-		Pubkey: []byte{},
-	}
-
-	_, err := s.SetFeeRecipientByPubkey(ctx, req)
-	require.ErrorContains(t, "not a valid bls public key", err)
-}
-
-func TestServer_SetGasLimit_InvalidFeeRecipient(t *testing.T) {
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-
-	byteval, err := hexutil.Decode("0xaf2e7ba294e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2c06bd3713cb442072ae591493")
-	require.NoError(t, err)
-
-	s := &Server{
-		validatorService: &client.ValidatorService{},
-	}
-
-	req := &ethpbservice.SetFeeRecipientByPubkeyRequest{
-		Pubkey: byteval,
-	}
-
-	_, err = s.SetFeeRecipientByPubkey(ctx, req)
-	require.ErrorContains(t, "Fee recipient is not a valid Ethereum address", err)
-}
-
-func TestServer_DeleteFeeRecipientByPubkey(t *testing.T) {
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-	byteval, err := hexutil.Decode("0xaf2e7ba294e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2c06bd3713cb442072ae591493")
-	require.NoError(t, err)
-	type want struct {
-		EthAddress string
-	}
-	tests := []struct {
-		name             string
-		proposerSettings *validatorserviceconfig.ProposerSettings
-		want             *want
-		wantErr          bool
-	}{
-		{
-			name: "Happy Path Test",
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(byteval): {
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x055Fb65722E7b2455012BFEBf6177F1D2e9738D5"),
-						},
-					},
-				},
-				DefaultConfig: &validatorserviceconfig.ProposerOption{
-					FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-						FeeRecipient: common.HexToAddress("0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9"),
-					},
-				},
-			},
-			want: &want{
-				EthAddress: common.HexToAddress("0x046Fb65722E7b2455012BFEBf6177F1D2e9738D9").Hex(),
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			m := &mock.MockValidator{}
-			err := m.SetProposerSettings(ctx, tt.proposerSettings)
-			require.NoError(t, err)
-			validatorDB := dbtest.SetupDB(t, [][fieldparams.BLSPubkeyLength]byte{})
-			vs, err := client.NewValidatorService(ctx, &client.Config{
-				Validator: m,
-				ValDB:     validatorDB,
-			})
-			require.NoError(t, err)
-			s := &Server{
-				validatorService: vs,
-				valDB:            validatorDB,
-			}
-			_, err = s.DeleteFeeRecipientByPubkey(ctx, &ethpbservice.PubkeyRequest{Pubkey: byteval})
-			require.NoError(t, err)
-
-			assert.Equal(t, true, s.validatorService.ProposerSettings().ProposeConfig[bytesutil.ToBytes48(byteval)].FeeRecipientConfig == nil)
-		})
-	}
-}
-
-func TestServer_DeleteFeeRecipientByPubkey_ValidatorServiceNil(t *testing.T) {
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-
-	s := &Server{}
-
-	_, err := s.DeleteFeeRecipientByPubkey(ctx, nil)
-	require.ErrorContains(t, "Validator service not ready", err)
-}
-
-func TestServer_DeleteFeeRecipientByPubkey_InvalidPubKey(t *testing.T) {
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-	s := &Server{
-		validatorService: &client.ValidatorService{},
-	}
-
-	req := &ethpbservice.PubkeyRequest{
-		Pubkey: []byte{},
-	}
-
-	_, err := s.DeleteFeeRecipientByPubkey(ctx, req)
-	require.ErrorContains(t, "not a valid bls public key", err)
 }
 
 func TestServer_GetGasLimit(t *testing.T) {
